switch to Go native types for iso8583 data structs

+++ issuer/models/account.go
@@ -8,20 +8,20 @@ import (
 var ErrInsufficientFunds = errors.New("insufficient funds")
 
 type CreateAccount struct {
-	Balance  int
+	Balance  int64
 	Currency string
 }
 
 type Account struct {
 	ID               string
-	AvailableBalance int
-	HoldBalance      int
+	AvailableBalance int64
+	HoldBalance      int64
 	Currency         string
 
 	mu sync.Mutex
 }
 
-func (a *Account) Hold(amount int) error {
+func (a *Account) Hold(amount int64) error {
 	a.mu.Lock()
 	defer a.mu.Unlock()
